gradual refactor of scripts
made a cabal script to automate creating and refreshing sandboxes.
refreshing the sandbox is much more heavy weight than refreshing profiles;
all dependencies need to be recompiled

made refresh.sh a little easier to work with. i think dividing it into a
script-based refresh and a makefile-based nuke would be a good idea. the
makefile would setup the system in a pristine state, and refresh could be
used when making minor changed to config files. maybe having a makefile
that sets up haskell and then delegates the remainder of the work to shake
if make proves insufficient, but for the time being i want to try make
again. i think the problem with using it before might have been using it
for refresh